extract localizedDispatcherRoute tests
